fix(#6831,#6832): one implementation of the openCypher SET clause
MERGE's ON CREATE SET / ON MATCH SET was served by a second, hand-maintained
copy of the SET switch. Both copies are fed by the SAME parser production
(visitSetClause), so both see every SET shape, and the two had drifted in
opposite directions:

- the MERGE copy read only item.getVariable() + item.getProperty(), so the
  dynamic-key form (SET n[$k] = 5) and the expression-target forms
  (SET (CASE WHEN ... THEN n END).x = 1) were silently dropped - the first
  even reaching MutableDocument.set(null, ...) and NPE'ing on a vertex type.
  It also skipped the property-value type check, so ON CREATE SET n.p = {a:1}
  stored a map where the identical stand-alone SET raises TypeError, and it
  evaluated items one by one, so ON MATCH SET n.x = n.y, n.y = n.x copied
  instead of swapping (#6831);
- the stand-alone copy required a java.util.Map right-hand side for
  SET n = y / SET n += y, so copying properties off another node or
  relationship - the documented way to clone an entity, and what the MERGE
  copy did correctly - was a silent no-op with no write, no error and no
  statistic (#6832).

Both are now SetClauseApplier, used by SetStep and MergeStep alike, so a SET
item behaves the same inside and outside a MERGE. A node or relationship
right-hand side contributes its properties (defensively copied: a
MutableDocument hands out a live view of its own map, which SET a = a would
otherwise read after emptying it), and anything that is neither an entity nor
a map is a TypeError rather than a discarded write.

The two call sites keep exactly the two differences that were deliberate,
named by the factory methods rather than by a divergent copy of the code:
forSetClause() reloads the write target to its latest committed version
(#5227), and forMergeAction() does not re-assert a value the record already
holds (#4474) - reloading there would pin the page and reintroduce the
write-conflict storm that fix exists to prevent.

The numeric-tolerant value comparison both steps needed moves to
CypherValues.equalValues() rather than being copied a third time.
